Adding replay_dir parameter to Api to allow testing without rate-limiting.
It's handy to be able to test some analysis logic without running afoul of
Twitter's rate-limiting. The staleness of the Twitter data often isn't
an issue when testing. Having a local cached copy of each
response that python-twitter could replay would avoid this problem.

This commit adds a new parameter to the Api class, called 'replay_dir'
and defaults it to None. If this parameter is specified, the Api
_RequestUrl(..) method will attempt to read the response from the
local cache dir before issuing a request to Twitter. If it does
make the request to Twitter, and replay_dir is specified, it will
save the response to disk for later replay.